run cargo clean before cargo build, just to make sure rust release builds are always fresh
The cargo build cache inside the dev VM is shared with the host OS, so any release builds done on the host OS could populate cargo's build cache. If the cargo inside the (very carefully curated CentOS) container decided to skip a build step because a cache was populated with build artifacts from the host OS, that could make the resulting binary a lot less portable than we'd like.