remove all manual block tab forwarding when launching jobs and rely on pyp configuration instead (to use the hidden tab copying system)
but will this even work? Say we import data and set the data path. Then create a preprocessing block, which copies the data path into the hidden data tab. Then we change the data path in the import block. Then try to run the preprocessing block. It won't see the new data path, right? Because there's no mechanism to update it.