implement particle versioning, switch everything to using unbinned coordinates for particles
New particles are now always read/written/stored in completely unbinned coordinates.
Old particles are stored in ... whatever they were stored in before. The website will make a best effort to convert them to unbinned coordinates before working with them based on the parameters of the job or session.
In the case of editing old manually-picked particles, the website will try to migrate them to the new unbinned version before making any changes.